style: 套用 dotnet format 全方案程式碼格式統一化
- 依據 .editorconfig 規則，對全方案 110 餘個 C# 程式碼檔案執行 dotnet format 統一格式。
- 涵蓋所有 Agent 代理程式、Admin 管理面板、Service 背景服務、Shared 共用程式庫與測試專案之排版標準化。
- 確保縮排、空行、括號位置與 using 引用排序符合一致性規範。

+++ Cyberarms.IntrusionDetection.Shared/TcpHeader.cs
@@ -30,7 +30,7 @@ public class TCPHeader
         uiAcknowledgementNumber = (uint)IPAddress.NetworkToHostOrder(binaryReader.ReadInt32());
         usDataOffsetAndFlags = (ushort)IPAddress.NetworkToHostOrder(binaryReader.ReadInt16());
         usWindow = (ushort)IPAddress.NetworkToHostOrder(binaryReader.ReadInt16());
-        sChecksum = (short)IPAddress.NetworkToHostOrder(binaryReader.ReadInt16());
+        sChecksum = IPAddress.NetworkToHostOrder(binaryReader.ReadInt16());
         usUrgentPointer = (ushort)IPAddress.NetworkToHostOrder(binaryReader.ReadInt16());
 
         byHeaderLength = (byte)(usDataOffsetAndFlags >> 12);
